+++ packages/alfa-aria/src/name/index.ts
@@ -1,0 +1,2 @@
+export * from "./name";
+export * from "./predicate";
