--- packages/alfa-aria/src/name/predicate.ts
@@ -1,1 +1,0 @@
-export * from "./predicate/has-value";
