+++ packages/alfa-aria/src/name/predicate/index.ts
@@ -1,0 +1,1 @@
+export * from "./has-value";
