+++ packages/alfa-aria/src/node/text.ts
@@ -2,7 +2,7 @@ import { Option } from "@siteimprove/alfa-option";
 
 import * as dom from "@siteimprove/alfa-dom";
 
-import { Name } from "../name";
+import { Name } from "../name/name";
 import { Node } from "../node";
 
 /**
