+++ packages/alfa-aria/tsconfig.json
@@ -24,8 +24,9 @@
     "src/dom/predicate/is-semantically-disabled.ts",
     "src/feature.ts",
     "src/index.ts",
-    "src/name.ts",
-    "src/name/predicate.ts",
+    "src/name/index.ts",
+    "src/name/name.ts",
+    "src/name/predicate/index.ts",
     "src/name/predicate/has-value.ts",
     "src/node.ts",
     "src/node/container.ts",
